Document workaround for syrupy module does not exist error
The syrupy pytest plugin parses its CLI arguments and tries to import
its `--snapshot-default-extension` very early. The module to import here
is test-only; it is not within this project's source folder. The module
is not on the Python path before pytest honors its own Python path via
pytest's `pythonpath` setting, nor before any conftest.py in this
project can modify the global path. We must manually bend our usage of
the `pytest` CLI to syrupy, and set the global Python path as early as
possible.