Remove grep removing minus symbol
Fixes pattern match highlighting.

The minus symbol's presence should be an exceptional case and therefore
handled case by case. The replacement initially was to ease pasting one
kind of transaction into a spreadsheet, withdrawals for paying my rent.
When those transactions were interspersed with other types of
transaction, there wouldn't be a mix of positive and negative numbers to
sum. However, the negative symbol could signal other things, for example
a refund. It should not be globally find/replaced. It should be handled
by manual review.